roro_sentinel/requirements.txt
# RORO SENTINEL CFD TRADING SYSTEM
# Python 3.11+ required

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Data handling
pyarrow>=12.0.0  # For parquet files

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3

# Optional: IBKR integration
# ib-insync>=0.9.86

# Optional: Additional data sources
# polygon-api-client>=1.12.0
# alpha-vantage>=2.3.1

# Optional: Alert channels
# discord.py>=2.3.0
# python-telegram-bot>=20.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging
colorlog>=6.7.0

# Type checking (optional)
# mypy>=1.5.0
